fix: Alias output validation code refactoring
